Add a columnWidthProvider, column headings and scrollbars
We want to know the maximum width of each column.  But assume that it only
needs to be calculated the first time.  There is a width limit anyway,
because the F_CMDLINE field in particular can be quite long.

The TableView scrolls under the headings.  It's useful to keep them
visible all the time.  But in the future TableView will have other options
like ListView does.