Make the columns sortable; reorder like "top"; QML timer; format data sizes
Reorder the columns: leftmost ones correspond to "top" output

Move the timer to QML; add a toolbar with "update" switch.
This allows us to turn updating off, e.g. to be able to
examine a row for a longer period.

Use QLocale::formattedDataSize for data sizes.
It's new in Qt 5.10 and makes this formatting much more
configurable and future-proof (supports really big prefixes,
will get revised if international standards are revised,
gets translated to various languages).