Sort numeric numerically; sort initially; add filter string feature
QSortFilterProxyModel already does that, but the QVariant returned
from data() needs to be a number, not a string.  But for DisplayRole
it should continue returning formatted strings.  So we need another
role. We add the Sort role to return a "sortable" data type, and call
setSortRole() to request QSortFilterProxyModel to use it.

Sort columns initially according to Qt::InitialSortOrderRole.
When sorting by memory or CPU usage or anything numeric,
put the row with the biggest number on top by default.
(Those are usually the "interesting" processes.)

Type-ahead search.