Get numeric sort working
QSortFilterProxyModel already does that, but the QVariant returned from data()
needs to be a number, not a string.  But for DisplayRole it should continue
returning formatted strings.  So we need another role. Too bad Qt::ItemDataRole
doesn't have one; it seems like an oversight considering that setSortRole()
has already been there for so long.  (or I'm missing something)