Use DelegateChooser to render different data types; bar graph for percent
DelegateChooser is in the Qt.labs.qmlmodels import and is also
new in Qt 5.12. It allows you to switch to different cell delegates
depending on various conditions: in this example, depending on
a "type" role provided by the table model.

Here I use a bar graph to show memory and CPU usage graphically,
and right-align all other numeric fields.